components/webhook/pyproject.toml
[project]
name = "ethitrust-webhook"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi==0.115.0",
    "uvicorn[standard]==0.30.6",
    "sqlalchemy[asyncio]==2.0.35",
    "asyncpg==0.29.0",
    "pydantic[email]==2.9.2",
    "pydantic-settings==2.5.2",
    "redis[asyncio]==5.1.0",
    "aio-pika==9.4.3",
    "grpcio==1.78.0",
    "grpcio-tools==1.78.0",
    "protobuf==6.31.1",
    "httpx==0.27.2",
    "PyJWT==2.9.0",
    "celery[redis]==5.4.0",
    "alembic>=1.13.0"
]
[project.optional-dependencies]
test = ["pytest==8.3.3","pytest-asyncio==0.24.0","httpx==0.27.2","aiosqlite==0.20.0"]
[tool.pytest.ini_options]
asyncio_mode = "auto"
